Water level sensor and buzzer
In this experiment we use water level sensor and buzzer to detect the water level and to notify us with a sound(made by buzzer) during water leakage. In part 2 of the LAB we have learnt how to use the touch pin and to turn ON the buzzer when a touch is encountered. The vedio uploaded is for the part 2.